refactor : FrameInfo dto 계층이동  - presentation layer에서 도메인 영역의 dto로 변경
